Crate casper_dao_utils

Re-exports

pub use casper_contract;

pub use casper_dao_macros;

pub use conversions::BytesConversion;

Modules

casper_env

Functions to interact with the host environment.

consts

Useful constants for common strings.

conversions

Types conversion.

cspr

Functions for interacting with purses.

definitions

Data structures describing contracts and events metadata.

math

Calculation utility functions.

Structs

ContractCall

A serializable data structure that represent a contract call.

IndexedMapping

Data structure for storing index-value pairs.

Iter

OrderedCollection iterator.

Mapping

Data structure for storing key-value pairs.

OrderedCollection

Data structure for storing indexed values.

TestEnv

CasperVM based testing environment.

Variable

Data structure for storing a single value.

VecMapping

Data structure for storing key-value pairs.

Enums

Address

An enum representing an AccountHash or a ContractPackageHash.

Error

All possible errors that can be raised by the utils crate.

ExecutionError

Possible execution errors.

Traits

Instanced

A trait that should be implemented by each module to instantiate it and set up the storage.

List

A collection acts like a list.

Set

A collection acts like a set.

TestContract

An interface of a contract to be tested against TestEnv.

Type Definitions

BlockTime

A value that represents the block time.

DocumentHash

The hash of a document sent to the system.