Module casper_dao_utils::consts

Useful constants for common strings.

Constants

ARG_ADDRESS

The name of address entry point argument.

ARG_AMOUNT

The name of amount entry point argument.

ARG_KEY

The name of key entry point argument.

ARG_OWNER

The name of owner entry point argument.

ARG_RECIPIENT

The name of recipient entry point argument.

ARG_TO

The name of to entry point argument.

ARG_TOKEN_ID

The name of token_id entry point argument.

ARG_VALUE

The name of value entry point argument.

BID_ESCROW_FORMAL_QUORUM_RATIO

How many VA’s are needed for an formal voting quorum.

BID_ESCROW_FORMAL_VOTING_TIME

Time for the formal part of the Bid Escrow voting.

BID_ESCROW_INFORMAL_QUORUM_RATIO

How many VA’s are needed for an informal voting quorum.

BID_ESCROW_INFORMAL_VOTING_TIME

Time for the informal part of the Bid Escrow voting.

BID_ESCROW_PAYMENT_RATIO

How much CSPR is sent to GovernanceWallet after the Job is finished

BID_ESCROW_WALLET_ADDRESS

An address of a multisig wallet (GovernanceWallet) of the DAO.

CONTRACT_MAIN_PURSE

DEFAULT_POLICING_RATE

Defines how many Reputation tokens are given to the VA’s for their community audit/vote on a work product.

DEFAULT_REPUTATION_SLASH

How much reputation of an Internal Worker is slashed after not completing a Job.

DISTRIBUTE_PAYMENT_TO_NON_VOTERS

Determines if the Payment for the Job should be distributed between all VA’s or only to those who voted.

EP_ADD_TO_WHITELIST

The name of add_to_whitelist entry point.

EP_BURN

The name of burn entry point.

EP_CHANGE_OWNERSHIP

The name of change_ownership entry point.

EP_GET

The name of get entry point.

EP_INIT

The name of init entry point.

EP_MINT

The name of mint entry point.

EP_REMOVE_FROM_WHITELIST

The name of remove_from_whitelist entry point.

EP_SET_OR_UPDATE

The name of set_or_update entry point.

EP_STAKE

The name of stake entry point.

EP_TRANSFER_FROM

The name of transfer_from entry point.

EP_UNSTAKE

The name of unstake entry point.

FIAT_CONVERSION_RATE_ADDRESS

An address of a contract that will return the conversion rate between Fiat and CSPR.

FORMAL_QUORUM_RATIO

How many VA’s are needed for an formal voting quorum.

FORMAL_VOTING_TIME

Time for the formal part of other votings

FORUM_KYC_REQUIRED

Defines if KYC is required to post on Forum.

INFORMAL_QUORUM_RATIO

How many VA’s are needed for an informal voting quorum.

INFORMAL_STAKE_REPUTATION

Tells if the Informal Voting should stake the reputation or only simulate it.

INFORMAL_VOTING_TIME

Time for the informal part of other votings

INTERNAL_AUCTION_TIME

The time of the Internal Auction.

LENGTH_SUFFIX

The key the collection length is stored under.

NAME_EVENTS

The key the contracts’ events are stored under.

POST_JOB_DOS_FEE

A DOS fee that the JobPoster needs to attach to the Post Job query. The value is the minimum amount of Fiat currency to be attached as CSPR using FiatConversionRate.

PUBLIC_AUCTION_TIME

The time of the External Auction.

REPUTATION_CONVERSION_RATE

Defines how much Reputation is minted for each unit of currency paid for Work.

TIME_BETWEEN_INFORMAL_AND_FORMAL_VOTING

Time between Informal and Formal Votings.

VA_BID_ACCEPTANCE_TIMEOUT

How much time the bid wait for the acceptance. After this time, the bid can be cancelled

VA_CAN_BID_ON_PUBLIC_AUCTION

Whether or not VA’s can take part in the Public Auction part of the Bidding process.

VOTING_CLEARNESS_DELTA

If the difference between 50/50 and result of the Informal Voting is bigger than the value, the time between votings should be doubled.

VOTING_IDS_ADDRESS

An address of a contract that generates a next voting id.

VOTING_START_AFTER_JOB_WORKER_SUBMISSION

Time between the worker job submission and the internal voting start.