Trait casper_dao_utils::conversions::BytesConversion

pub trait BytesConversion: Sized {
    fn convert_to_bytes(&self) -> Result<Bytes, Error>;
    fn convert_from_bytes(bytes: Bytes) -> Result<Self, Error>;
}

Error-safe conversion from/to bytes.

Required Methods

fn convert_to_bytes(&self) -> Result<Bytes, Error>

Converts the struct to Bytes or returns an Error.

fn convert_from_bytes(bytes: Bytes) -> Result<Self, Error>

Converts Bytes to a struct or returns an Error.