Enum casper_dao_utils::ExecutionError

pub enum ExecutionError {

    Interpreter(String),
    Storage(Error),
    BytesRepr(Error),
    NamedKeyNotFound(String),
    KeyNotFound(Key),
    AccountNotFound(Key),
    TypeMismatch(TypeMismatch),
    InvalidAccess {
        required: AccessRights,
    },
    ForgedReference(URef),
    URefNotFound(URef),
    FunctionNotFound(String),
    ParityWasm(Error),
    WasmOptimizer,
    GasLimit,
    Ret(Vec<URef, Global>),
    Resolver(ResolverError),
    Revert(ApiError),
    AddKeyFailure(AddKeyFailure),
    RemoveKeyFailure(RemoveKeyFailure),
    UpdateKeyFailure(UpdateKeyFailure),
    SetThresholdFailure(SetThresholdFailure),
    SystemContract(Error),
    DeploymentAuthorizationFailure,
    ExpectedReturnValue,
    UnexpectedReturnValue,
    InvalidContext,
    IncompatibleProtocolMajorVersion {
        expected: u32,
        actual: u32,
    },
    CLValue(CLValueError),
    HostBufferEmpty,
    UnsupportedWasmStart,
    NoActiveContractVersions(ContractPackageHash),
    InvalidContractVersion(ContractVersionKey),
    NoSuchMethod(String),
    WasmPreprocessing(PreprocessingError),
    KeyIsNotAURef(Key),
    UnexpectedStoredValueVariant,
    LockedContract(ContractPackageHash),
    InvalidContractPackage(ContractPackageHash),
    InvalidContract(ContractHash),
    InvalidContractWasm(ContractWasmHash),
    MissingArgument {
        name: String,
    },
    DictionaryItemKeyExceedsLength,
    MissingSystemContractRegistry,
    MissingSystemContractHash(String),
    RuntimeStackOverflow,
    ValueTooLarge,
    MissingRuntimeStack,

}

Possible execution errors.

Variants

Interpreter(String)

WASM interpreter error.

Storage(Error)

Storage error.

BytesRepr(Error)

Failed to (de)serialize bytes.

NamedKeyNotFound(String)

Unable to find named key.

KeyNotFound(Key)

Unable to find a key.

AccountNotFound(Key)

Unable to find an account.

TypeMismatch(TypeMismatch)

Type mismatch error.

InvalidAccess

Fields

required: AccessRights

Required access rights of the operation.

Invalid access.

ForgedReference(URef)

Forged reference error.

URefNotFound(URef)

Unable to find a URef.

FunctionNotFound(String)

Unable to find a function.

ParityWasm(Error)

Parity WASM error.

WasmOptimizer

Error optimizing WASM.

GasLimit

Execution exceeded the gas limit.

Ret(Vec<URef, Global>)

A stored smart contract incorrectly called a ret function.

Resolver(ResolverError)

Error using WASM host function resolver.

Revert(ApiError)

Reverts execution with a provided status

AddKeyFailure(AddKeyFailure)

Error adding an associated key.

RemoveKeyFailure(RemoveKeyFailure)

Error removing an associated key.

UpdateKeyFailure(UpdateKeyFailure)

Error updating an associated key.

SetThresholdFailure(SetThresholdFailure)

Error setting threshold on associated key.

SystemContract(Error)

Error executing system contract.

DeploymentAuthorizationFailure

Weight of all used associated keys does not meet account’s deploy threshold.

ExpectedReturnValue

Host buffer expected a value to be present.

UnexpectedReturnValue

Host buffer was not expected to contain a value.

InvalidContext

Error calling a host function in a wrong context.

IncompatibleProtocolMajorVersion

Fields

expected: u32

Expected major version.

actual: u32

Actual major version supplied.

Unable to execute a deploy with invalid major protocol version.

CLValue(CLValueError)

Error converting a CLValue.

HostBufferEmpty

Unable to access host buffer.

UnsupportedWasmStart

WASM bytes contains an unsupported “start” section.

NoActiveContractVersions(ContractPackageHash)

Contract package has no active contract versions.

InvalidContractVersion(ContractVersionKey)

Invalid contract version supplied.

NoSuchMethod(String)

Contract does not have specified entry point.

WasmPreprocessing(PreprocessingError)

Error processing WASM bytes.

KeyIsNotAURef(Key)

Unable to convert a Key into an URef.

UnexpectedStoredValueVariant

Unexpected variant of a stored value.

LockedContract(ContractPackageHash)

Error upgrading a locked contract package.

InvalidContractPackage(ContractPackageHash)

Unable to find a contract package by a specified hash address.

InvalidContract(ContractHash)

Unable to find a contract by a specified hash address.

InvalidContractWasm(ContractWasmHash)

Unable to find the WASM bytes specified by a hash address.

MissingArgument

Fields

name: String

Name of the required argument.

Error calling a smart contract with a missing argument.

DictionaryItemKeyExceedsLength

Error writing a dictionary item key which exceeded maximum allowed length.

MissingSystemContractRegistry

Missing system contract registry.

MissingSystemContractHash(String)

Missing system contract hash.

RuntimeStackOverflow

An attempt to push to the runtime stack which is already at the maximum height.

ValueTooLarge

An attempt to write a value to global state where its serialized size is too large.

MissingRuntimeStack

The runtime stack is None.

Implementations

impl Error

pub fn type_mismatch(expected: CLType, found: CLType) -> Error

Returns new type mismatch error.