Trait casper_dao_utils::Instanced

pub trait Instanced {
    fn instance(namespace: &str) -> Self;
}

A trait that should be implemented by each module to instantiate it and set up the storage.

Required Methods

fn instance(namespace: &str) -> Self

Instantiates the module.